Readded names to account report + cleanup
