+++ app/preload/files.ts
@@ -1,8 +1,7 @@
-import AdmZip from 'adm-zip';
 import { ipcRenderer, OpenDialogOptions, SaveDialogOptions } from 'electron';
 import fs from 'fs';
 import { getDatabaseFilename } from '~/database/knexfile';
-import { SaveFileData, FileMethods } from '~/preload/preloadTypes';
+import { FileMethods } from '~/preload/preloadTypes';
 import ipcCommands from '~/constants/ipcCommands.json';
 
 /**
@@ -29,35 +28,6 @@ async function saveFile(filepath: string, data: string | Buffer) {
     });
 }
 
-async function saveZipFile(files: SaveFileData[]) {
-    const zip = new AdmZip();
-    for (const file of files) {
-        zip.addFile(file.filename, file.data);
-    }
-
-    const opts = {
-        title: 'Save Account Reports',
-        defaultPath: 'reports.zip',
-        filters: [{ name: 'zip', extensions: ['zip'] }],
-    };
-
-    const saveFileDialog = await ipcRenderer.invoke(
-        ipcCommands.saveFileDialog,
-        opts
-    );
-    if (saveFileDialog.canceled) {
-        return false;
-    }
-
-    if (!saveFileDialog.filePath) {
-        return Promise.reject(
-            new Error('No file path was selected by the user.')
-        );
-    }
-
-    return saveFile(saveFileDialog.filePath, zip.toBuffer());
-}
-
 const exposedMethods: FileMethods = {
     databaseExists,
     // Provides access to save file dialog from renderer processes.
@@ -73,7 +43,6 @@ const exposedMethods: FileMethods = {
             return false;
         }
     },
-    saveZipFileDialog: saveZipFile,
 };
 
 export default exposedMethods;
