+++ app/utils/FileHelper.ts
@@ -22,34 +22,10 @@ export async function openFileDestination(
 }
 
 /**
- * Opens a 'save file' prompt where the user can select where to save a file, and writes
- * the data to that destination.
- * @param data the string or buffer to save to a file
- * @param title the title of the save file window
- * @param defaultPath the default path that the save window starts on. Can be used just to set default file name.
- * @return the promise resolves true when the data has been written to file. if the result is false, then the user cancelled.
+ * Opens a 'save file' prompt where the user can select where to save a file, and returns the chosen destination.
+ * @param opts options for the dialog
+ * @return returns the chosen destination, or undefined if the dialog was canceled
  */
-export default async function saveFile(
-    data: Buffer | string,
-    opts: Electron.SaveDialogOptions
-): Promise<boolean> {
-    const saveFileDialog: Electron.SaveDialogReturnValue = await window.files.saveFileDialog(
-        opts
-    );
-
-    if (saveFileDialog.canceled) {
-        return false;
-    }
-
-    if (!saveFileDialog.filePath) {
-        return Promise.reject(
-            new Error('No file path was selected by the user.')
-        );
-    }
-
-    return window.files.saveFile(saveFileDialog.filePath, data);
-}
-
 export async function chooseFileDestination(
     opts: Electron.SaveDialogOptions
 ): Promise<string | undefined> {
@@ -62,12 +38,30 @@ export async function chooseFileDestination(
     }
 
     if (!saveFileDialog.filePath) {
-        return Promise.reject(
-            new Error('No file path was selected by the user.')
-        );
+        throw new Error('No file path was selected by the user.');
     }
 
     return saveFileDialog.filePath;
+}
+
+/**
+ * Opens a 'save file' prompt where the user can select where to save a file, and writes
+ * the data to that destination.
+ * @param data the string or buffer to save to a file
+ * @param opts options for the dialog
+ * @return the promise resolves true when the data has been written to file. if the result is false, then the user cancelled.
+ */
+export default async function saveFile(
+    data: Buffer | string,
+    opts: Electron.SaveDialogOptions
+): Promise<boolean> {
+    const filePath = await chooseFileDestination(opts);
+
+    if (!filePath) {
+        return false;
+    }
+
+    return window.files.saveFile(filePath, data);
 }
 
 /**
