+++ app/utils/basicHelpers.ts
@@ -159,10 +159,6 @@ export function max(first: bigint, second: bigint) {
     return first > second ? first : second;
 }
 
-export function min(first: bigint, second: bigint) {
-    return first < second ? first : second;
-}
-
 export function isASCII(value: string) {
     // eslint-disable-next-line no-control-regex
     return /[^\u0000-\u007f]/.test(value);
